Economy + Dashboard
Added a fixed economy that will be changed later on to match the database, made it so that you have a drop down section so that you can enter it properly. Put a server.ts, doesn't have to be changed to tsx. Added a functional database and a connection to all, also installed the required packages